refactor BWS express app routes
